feat: make distribution compatible with deno (#21)
* Add esbuild plugin and improve JSON imports

Included the `esbuild-plugin-file-path-extensions` dependency and adjusted esbuild configuration to use the plugin.

* Update module extensions and improve build scripts

Renamed module file extensions to .cjs and .mjs for better clarity and added new npm scripts for versioning and publishing. Refactored and moved esbuild script to the scripts directory and added new esbuild plugin to handle import assertions.

* Add Deno usage instructions to README

Introduced a new section in the README for running the package with Deno. This includes information on necessary permissions and a workaround for dependencies needing environment permissions.